Refactor items
Otherwise silver will get mad.
