+++ mFriesen-S2TextBasedRPG/Scripts/EntityThings/Pickup.cs
@@ -31,90 +31,4 @@ namespace mFriesen_S2TextBasedRPG
             return e;
         }
     }
-
-    class RestorationPickup : Pickup
-    {
-        // References and values
-        public restorationType rType;
-
-        public int rValue;
-
-        public RestorationPickup(Vector2 position, restorationType rType, int value)
-        {
-            this.position = position;
-            this.rType = rType; rValue = value;
-
-            SetDefaultValues();
-        }
-
-        // Do not override SetDefaultValues.
-
-        public override Entity DeepClone()
-        {
-            RestorationPickup r = (RestorationPickup)MemberwiseClone();
-            r.position = position.Clone();
-            r.displayTile = displayTile.Clone();
-            return r;
-        }
-    }
-
-    class ItemPickup : Pickup
-    {
-        public Item item { get; private set; }
-        public ItemPickup(Vector2 position, Item item)
-        {
-            this.position = position;
-            this.item = item;
-
-            SetDefaultValues();
-        }
-
-        protected override void SetDefaultValues()
-        {
-            // set tile
-            displayTile = new Tile();
-            displayTile.displayChar = 'i';
-            displayTile.bg = ConsoleColor.White;
-            displayTile.fg = ConsoleColor.Blue;
-            displayTile.hazard = Hazard.none;
-        }
-
-        public override Entity DeepClone()
-        {
-            ItemPickup e = (ItemPickup)MemberwiseClone();
-            e.position = position.Clone();
-            e.displayTile = displayTile.Clone();
-            e.item = item;
-            return e;
-        }
-    }
-
-    class EffectPickup : Pickup
-    {
-        public StatusEffect effect { get; private set; }
-        public EffectPickup(Vector2 position, StatusEffect effect)
-        {
-            this.effect = effect;
-            this.position = position;
-        }
-
-        protected override void SetDefaultValues()
-        {
-            // set tile
-            displayTile = new Tile();
-            displayTile.displayChar = 'e';
-            displayTile.bg = ConsoleColor.White;
-            displayTile.fg = ConsoleColor.DarkRed;
-            displayTile.hazard = Hazard.none;
-        }
-
-        public override Entity DeepClone()
-        {
-            EffectPickup e = (EffectPickup)MemberwiseClone();
-            e.position = position.Clone();
-            e.displayTile = displayTile.Clone();
-            e.effect = effect;
-            return e;
-        }
-    }
 }
